refactor: share PostForm
